pyproject.toml
[project]
name = "cram-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Kolja Sam",email = "github@koljapluemer.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]

[tool.poetry]
packages = [{include = "cram_backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
